refactor(parser): Remove AstParenthesizedExpression
The Parenthesized Expression is now parsed to a simple AstExpression
node. Because of this, the following functions have had their return
types changed to AstExpression as well:

- parseConditionalExpression()
- parsePrimaryExpression()

Also, replace AstPrimaryExpression with AstIntegralExpression. This new
class is the parent of all primary expressions except the Parenthesized
Expression.

When checking if the reference argument in a reference function call is
actually a primary expression, check if either is true:

- The expression is an AstIntegralExpression, or
- The expression is a Parenthesized Expression
  (i.e. the first argument is an opening parenthesis)