feat: Add 'type' property to all AST nodes
This property will be used as a tag that discriminates AST nodes.

Note: The 'type' property must be directly on at least one subtype in a
union for TypeScript to correctly discriminate them.
Since it would be awkward to have only one AST class assign to
`this.type`, I made _all_ AST classes assign to `this.type` instead of
`Class.prototype.type`.