feat: Use d2calc.umd.js as pkg.main
Because older versions of Node.js (10.x) do not recognize the '.cjs'
file extension when running with the '--experimental-modules' flag,
we need to provide a script with '.js' extension in the "main" field.
Therefore, generate a copy of the UMD bundle with the '.js' extension
and use it as the "main".

Newer versions of Node.js (12.x) won't be affected, as they will ignore
the "main" field, and instead use the CJS or ESM scripts specified under
the "exports" field.