fix!(interp): Only use own properties in env
Previously, when interpret() read in identifiers, numeric functions, and
reference functions from InterpreterEnvironment, it also used properties
inherited through the prototype chain. This caused unexpected properties
(such as those in the Object.prototype) to leak into the environment.
To avoid this, only use properties directly owned by the fields in
InterpreterEnvironment.

Note that this restriction does not apply to the top-level
InterpreterEnvironment object itself. This is OK, since interpret() only
reads known property names from it.