fix(interp): Object.prototype props don't pollute cache
Previously, CachedInterpreter used a plain object to cache ASTs. This
caused CachedInterpreter.interpret() to confuse properties inherited
through Object.prototype as "cached" ASTs.

To avoid this, CachedInterpreter now uses a Map instead of a plain
object. This is simpler than using Object.prototype.hasOwnProperty().

(Note: This method could not be applied to the interpret() fix, because
InterpreterEnvironment is a part of the public API, and it's easier for
the user to supply a plain object rather than a Map.)

Also update CachedInterpreter tests to use the Map.