chore(ts): Rename to jsconfig.build.json
Since this file is used to emit type declarations from JavaScript files,
not TypeScript, the "tsconfig.*.json" file name was misleading.
Previously, however, I was forced to use "tsconfig.*.json" because VS
Code did not recognize "jsconfig.*.json" as JSONC (JSON with Comments).
This was fixed in https://github.com/microsoft/vscode/pull/103360.
Now that the change is shipped in VS Code 1.48, we can safely rename the
file to jsconfig.build.json.

Note: VS Code complains about the "declarationMap" property, because the
JSON schema for jsconfig.json--provided by JSON Schema Store
(https://www.schemastore.org/json/)--does not have it. This seems to be
an oversight; I will submit a PR to remedy this.